chore: Slightly restructured showcase post
The showcase index page will include the first lines of the showcase
post.

In this case, most of the info was in the image caption.
The body itself was about the picture being posted with permission.

So I moved some of the description to the post body as that looks
better on the index page.