Fix: Double 'the' and more context in hyperlink
There was a double `the` and since I was add it, I expanded the link text a bit to make it more meaningful.

For robots (like a Google crawler) or screen readers, a link that just say `andrew` is not very informative.  
A link that says `the Instagram feed of @sewandrew` tells you exactly what to expect when you click it.