Update en.md
Hi Joost,

I started this the day after the release, as I thought you'd like to get the typos out before it had been up for too long, but I was still having trouble with the process. I think I've got it now. 

One question: In the penultimate paragraph *Speaking of documentation*, should the link for **https://freesewing.dev/** really go to **https://freesewing.dev/tutorials/pattern-design/**, where that switch actually is? (And should the switch be labelled?)

I also have a few other editing/proofing points I want discuss with you. I tried to attach a document to a DM, but it didn't seem to be possible. What's the best way to send a page or two of text to a single Discord user?