Update en.md
*“Behind the seams: An FIXME”*    FIXME is recondite,  so won’t mean anything to non-coders, who are two of out three user groups. Together with “An”, which is an unusual spelling for a name, it’s a bit of a barrier to this article. 

“Joost sent me an e-mail with a link to FreeSewing’s Discord”    Possibly consider saying what Discord is, some users of the site may not know.

*Anything you dislike about sewing?  Honestly? I dislike most of the process other than actually seaming the pieces together.* Was this meant to be “sewing the pieces together”?

*“Gender is an anti-pattern.  An anti-pattern is a term used in technology to label a common approach that is counter-productive.*
You’ve defined “anti-pattern” within technology, which is helpful, but it might be even better to stress that it bears no relation to a sewing pattern, which is what most readers will understand as the meaning of that word.

Does the line configuration matter in the github issue? When I copied it out and back all the line-ends changed. But I see that the line in  Behind the seams with the hyperlink resolves OK, so does that mean they all do?