Update en.md
Just having a play around with the format to try and allign the images to the right.
No idea what this would like with the <Note> <Tip> <Warning> etc. formatting. Probably would put those at the bottom so they are under the image but have not done that here.
This is not nessecarily an update for Aaron. I just used these instructions as they are complete.  I thought it would be an idea to just put this out there to see what peeps thoughts are to format instructions like this so I know when writing instructions from scratch what people prefer.