Bold issue. Holmes Fabric Options
For some reason in the notes "Medium Firm Interfacing" and "Tarlantan" were not appearing bold on the website but were appearing bolod in the preview. I have changed them to the __ markdown instead to see if this solves the issue. If not I'm not sure why as on the other pages the * worked fine.